+++ src/include/systick.h
@@ -1,6 +1,8 @@
 #pragma once
 #include <stdint.h>
 
+#define SYSTICK_HZ 100 // Default SysTick frequency in Hz
+
 /**
  * @brief Initializes the SysTick timer to generate interrupts at a given frequency.
  *